CNAM-418: Reorganize files.
