+++ src/main/scala/fr/polytechnique/cmap/cnam/etl/transformers/interaction/Period.scala
@@ -1,0 +1,60 @@
+// License: BSD 3 clause
+
+package fr.polytechnique.cmap.cnam.etl.transformers.interaction
+
+import java.sql.Timestamp
+import fr.polytechnique.cmap.cnam.util.functions._
+
+case class Period(start: Timestamp, end: Timestamp) extends Remainable[Period]{
+  self =>
+
+  def & (other: Period): Boolean = {
+    (self.start.compareTo(other.start), self.end.compareTo(other.end)) match {
+      case (0, _) => true
+      case (_, 0) => true
+      case (1, -1) => true
+      case (-1, 1) => true
+      case (1, 1) => self.start.compareTo(other.end) match {
+        case 0 => false
+        case 1 => false
+        case -1 => true
+      }
+      case (-1, -1) => self.end.compareTo(other.start) match {
+        case 0 => false
+        case -1 => false
+        case 1 => true
+      }
+    }
+  }
+
+  def merge(other: Period): Period = Period(max(self.start, other.start), min(self.end, other.end))
+
+  def - (other: Period): RemainingPeriod[Period] = {
+    (self.start.compareTo(other.start), self.end.compareTo(other.end)) match {
+      case (0, 0) => NullRemainingPeriod
+      case (0, 1) => RightRemainingPeriod(Period(other.end, self.end))
+      case (0, -1) => NullRemainingPeriod
+      case (1, 0) => NullRemainingPeriod
+      case (1, 1) => self.start.compareTo(other.end) match {
+        case 0 => RightRemainingPeriod(self)
+        case 1 => RightRemainingPeriod(self)
+        case -1 => RightRemainingPeriod(Period(other.end, self.end))
+      }
+      case (1, -1) => NullRemainingPeriod
+      case (-1, 0) => LeftRemainingPeriod(Period(self.start, other.start))
+      case (-1, 1) => DisjointedRemainingPeriod(
+        LeftRemainingPeriod(Period(self.start, other.start)),
+        RightRemainingPeriod(Period(other.end, self.end))
+      )
+      case (-1, -1) => self.end.compareTo(other.start) match {
+        case 0 => LeftRemainingPeriod(self)
+        case -1 => LeftRemainingPeriod(self)
+        case 1 => LeftRemainingPeriod(Period(self.start, other.start))
+      }
+    }
+  }
+}
+
+object Period extends Ordering[Period] {
+  override def compare(x: Period, y: Period): Int = x.start.compareTo(y.start)
+}
