+++ src/main/scala/fr/polytechnique/cmap/cnam/etl/transformers/interaction/Remainable.scala
@@ -1,0 +1,7 @@
+// License: BSD 3 clause
+
+package fr.polytechnique.cmap.cnam.etl.transformers.interaction
+
+trait Remainable[A] {
+  def - (other: A): RemainingPeriod[A]
+}
